Better Walking

+++ src/player.h
@@ -2,10 +2,10 @@
 
 // player struct
 struct Player{
-  int xPos;
-  int yPos;
-  int dY;
-  int dX;
+  float xPos;
+  float yPos;
+ float dY;
+  float dX;
 
   void (*update)(struct Player *);
   void (*jump)(struct Player *);
